+++ src/interfaces/INodeData.ts
@@ -8,14 +8,14 @@
  *    and what (if any) child nodes need to be fetched.
  */
 
-import ObjectTypeDefinition from '../documents/ObjectTypeDefinition';
-import ObjectAttributeGroup from '../documents/ObjectAttributeGroup';
 import ObjectAttributeDefinition from '../documents/ObjectAttributeDefinition';
+import ObjectAttributeGroup from '../documents/ObjectAttributeGroup';
 import ObjectAttributeValueDefinition from '../documents/ObjectAttributeValueDefinition';
+import ObjectTypeDefinition from '../documents/ObjectTypeDefinition';
 import Site from '../documents/Site';
 
 /** @interface INodeData */
-export default interface INodeData {
+export interface INodeData {
   parentId: string;
   groupAttribute?: string;
   objectAttributeDefinition?: ObjectAttributeDefinition;
@@ -27,6 +27,14 @@ export default interface INodeData {
   baseNodeName?: string;
   nodeValue?: string | number;
   displayDescription?: string;
-  rootTree?:string;
+  rootTree?: string;
   site?: Site;
+  preferenceValue?: IPreferenceValue;
+}
+
+/** @interface IPreferenceValue */
+export interface IPreferenceValue {
+  id: string;
+  type: string;
+  enumValues?: string[] | number[];
 }